+++ lib/hubspot/discovery/events/send/api/custom_event_data_api.rb
@@ -2,9 +2,9 @@ require_relative '../../../base_api_client'
 
 module Hubspot
   module Discovery
-    module Files
-      module Files
-        class FilesApi
+    module Events
+      module Send
+        class CustomEventDataApi
           include Hubspot::Discovery::BaseApiClient
         end
       end