+++ lib/hubspot/discovery/files/api/folders_api.rb
@@ -1,0 +1,11 @@
+require_relative '../../base_api_client'
+
+module Hubspot
+  module Discovery
+    module Files
+      class FoldersApi
+        include Hubspot::Discovery::BaseApiClient
+      end
+    end
+  end
+end