+++ lib/hubspot/discovery/marketing/events/api/basic_api.rb
@@ -2,9 +2,9 @@ require_relative '../../../base_api_client'
 
 module Hubspot
   module Discovery
-    module Files
-      module Files
-        class FoldersApi
+    module Marketing
+      module Events
+        class BasicApi
           include Hubspot::Discovery::BaseApiClient
         end
       end