+++ lib/hubspot/discovery/marketing/events/api/subscriber_state_changes_api.rb
@@ -2,9 +2,9 @@ require_relative '../../../base_api_client'
 
 module Hubspot
   module Discovery
-    module Events
-      module Send
-        class BehavioralEventsTrackingApi
+    module Marketing
+      module Events
+        class SubscriberStateChangesApi
           include Hubspot::Discovery::BaseApiClient
         end
       end